requirements.txt
openai-whisper>=20231117
mlx-whisper>=0.1.0
openai>=1.0.0
tiktoken>=0.5.0
pytubefix>=6.0.0
pandas>=1.5.0
markdown>=3.4.0
tqdm>=4.65.0